requirements.txt
ai2thor>=3.3.4
torch==1.9.0
torchvision==0.10.0
matplotlib==3.3.4
tqdm==4.62.2
pyyaml==5.4.1
packaging==21.3
scipy==1.8.0
